Moved read/write clusters to parsing
